feat: made custom hook for signUpFirstStep
